FastAPI/requirements.txt
fastapi[all]
uvicorn
google-cloud-texttospeech
python-dotenv
